+++ chess-website/server/dbActions/findUser.js
@@ -1,4 +1,4 @@
-const signUpTemplateCopy = require('../models/SignUpModels')
+const signUpTemplateCopy = require('../models/SignUp')
 
 module.exports = async (email) => {
     return await signUpTemplateCopy.findOne({"email": email});
